Reorganize prompts
